// ui/src/generated/ModSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DeviceId } from "./DeviceId";
import type { ModulatorId } from "./ModulatorId";

/**
 * A modulation source.
 */
export type ModSource = { "type": "Modulator", modulator: ModulatorId, } | { "type": "Macro", rack: DeviceId, index: number, };